requirements.txt
# Add Python Dependencies
google-cloud-compute
google-api-core